fix compiler warnings

+++ src/Dataloader.cpp
@@ -120,7 +120,7 @@ Dataset DataLoader::load(const std::string &path, const std::string &label_col,
         for (size_t j = 0; j < n_features; j++) {
             X_train.at(i, j) = X.at(src, j);
         }
-        for (int k = 0; k < n_classes; k++) {
+        for (size_t k = 0; k < n_classes; k++) {
             Y_train.at(i, k) = Y.at(src, k);
         }
     }
